Move Networking package test code
