+++ PublishSubscribeBroker.csproj
@@ -3,6 +3,7 @@
   <PropertyGroup>
     <OutputType>Exe</OutputType>
     <TargetFramework>netcoreapp3.1</TargetFramework>
+    <StartupObject>PublishSubscribeBroker.Program</StartupObject>
   </PropertyGroup>
 
 </Project>
